@@ -3,84 +3,41 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "bootstrap": "^4.5.3",
-    "camelcase": "^5.2.0",
-    "classnames": "^2.2.6",
-    "fuse.js": "^3.6.1",
-    "lodash": "^4.17.20",
+    "bootstrap": "^5.2.2",
+    "camelcase": "^7.0.0",
+    "classnames": "^2.3.2",
+    "firebase": "^9.14.0",
+    "fuse.js": "^6.6.2",
+    "lodash": "^4.17.21",
     "open-iconic": "^1.1.1",
-    "react": "^16.14.0",
-    "react-bootstrap": "^1.4.0",
-    "react-dom": "^16.14.0",
-    "react-router-dom": "^5.2.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.6.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.3",
     "safe-access": "^0.1.0",
-    "uuid": "^3.4.0"
+    "uuid": "^9.0.0"
   },
   "devDependencies": {
-    "@babel/core": "7.*",
-    "@svgr/webpack": "4.3.2",
-    "@typescript-eslint/eslint-plugin": "^2.34.0",
-    "@typescript-eslint/parser": "^2.34.0",
-    "babel-eslint": "10.0.3",
-    "babel-jest": "^24.9.0",
-    "babel-loader": "8.0.6",
-    "babel-plugin-named-asset-import": "^0.3.7",
-    "babel-preset-react-app": "^9.1.2",
-    "case-sensitive-paths-webpack-plugin": "2.2.0",
-    "css-loader": "2.1.1",
-    "dotenv": "6.2.0",
-    "dotenv-expand": "5.1.0",
-    "eslint": "^6.8.0",
-    "eslint-config-react-app": "^5.2.1",
-    "eslint-loader": "3.0.2",
-    "eslint-plugin-flowtype": "3.13.0",
-    "eslint-plugin-import": "2.18.2",
-    "eslint-plugin-jsx-a11y": "6.2.3",
-    "eslint-plugin-react": "7.14.3",
-    "eslint-plugin-react-hooks": "^1.6.1",
-    "file-loader": "3.0.1",
-    "firebase": "^7.24.0",
-    "fs-extra": "7.0.1",
-    "html-webpack-plugin": "4.0.0-beta.5",
-    "is-wsl": "^1.1.0",
-    "jest": "24.9.0",
-    "jest-environment-jsdom-fourteen": "0.1.0",
-    "jest-resolve": "24.9.0",
-    "jest-watch-typeahead": "0.4.0",
-    "metascraper": "^5.14.22",
-    "mini-css-extract-plugin": "0.8.0",
-    "node-sass": "^4.14.1",
-    "optimize-css-assets-webpack-plugin": "5.0.3",
-    "pnp-webpack-plugin": "1.5.0",
-    "postcss-flexbugs-fixes": "4.1.0",
-    "postcss-loader": "3.0.0",
-    "postcss-normalize": "7.0.1",
-    "postcss-preset-env": "6.7.0",
-    "postcss-safe-parser": "4.0.1",
-    "react-app-polyfill": "^1.0.6",
-    "react-dev-utils": "^9.1.0",
-    "resolve": "1.12.0",
-    "resolve-url-loader": "^3.1.2",
-    "sass-loader": "7.2.0",
-    "semver": "6.3.0",
-    "style-loader": "1.0.0",
-    "terser-webpack-plugin": "^1.4.1",
-    "ts-pnp": "1.1.4",
-    "url-loader": "2.1.0",
-    "webpack": "4.41.0",
-    "webpack-dev-server": "^3.11.0",
-    "webpack-manifest-plugin": "2.1.1",
-    "workbox-webpack-plugin": "4.3.1"
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "eslint-config-react-app": "^7.0.1",
+    "react-scripts": "5.0.1",
+    "sass": "^1.56.1"
   },
   "scripts": {
-    "start": "node scripts/start.js",
-    "build": "node scripts/build.js",
-    "test": "node scripts/test.js",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
     "deploy": "firebase deploy",
     "predeploy": "npm run build"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
